_updateStats() Bug
So... I've been getting this weird bug 'expected floating-point number but got "PY_VAR8"', and I'm really not sure what that means.  I did a decent amount of refactoring, and it's improving, but I can't continue to debug without figuring out this line (447 & 450).  I provided documentation on the function.  If there's anyway you could check it out, that would be amazing!  Otherwise, I'll look at it again when I'm fresher.  :)