RBs_Bound/RefineDB_Error_Fixed
The bug from earlier about not being able to refine the database is fixed now.  It turns out I was resetting the search entries to empty strings before filtering the database.  I also fixed the error about the currentQuestion flashing English and Chinese consecutively every tenth of a second when in the 'Zh' language mode and I added the swithLangtoEn() and switchLangtoZh() functions.